Removed webpack-config
